pyproject.toml
[tool.poetry]
name = "feedbackfrontiers"
version = "0.1.0"
description = ""
authors = ["MeMartijn <martijn.schouten@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
pandas = "^2.1.2"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
tqdm = "^4.66.1"
markdownify = "^0.11.6"
langchain = "^0.0.333"
openai = "^1.2.2"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.1"
faiss-cpu = "^1.7.4"
streamlit = "^1.28.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
